Created README.md file and added basic data to it
In this commit, we have added the README.md file with some basic information inside it. The information added will be updated soon enough to a more clean, proper, and attractive documentation. Right now, the file is filled with only those information that can give anyone the very basic introduction of the repository / project.