Added the README.md file to the 'russian-basics' folder
In this commit, we add the README.md file to the 'basics' section of the Russian course folder. This doc file contains the introduction and slight start for the Russian basics course / documentation. The file is not fully organized neither it contains very good form of information, withing a while, more proper information (a proper documentation) will be provided with a proper update.